changed deleting images, fixed errors with positioning and reseting

+++ php/actions/exit.php
@@ -1,8 +1,10 @@
 <?php 
 	if ( isset($_GET['exit']) ){
 		session_start();
-		unlink($_SESSION['pathToMainFile']);
-		unlink($_SESSION['pathToMarkFile']);
+		if(isset($_SESSION['pathToMainFile']))
+			unlink($_SESSION['pathToMainFile']);
+		if(isset($_SESSION['pathToMarkFile']))
+			unlink($_SESSION['pathToMarkFile']);
 		session_destroy();
 	}
 ?>
